Revert "Split out ValidateModuleTopographyTask (#1133)"
This reverts commit 3644970d0cc6a5359826b64a37b51243001fc253.
